Standalone mode proposal (#46)
* Add stand alone configuration

* Add flushing for memory storage

* Add repositories for memory storage

* Refactoring for elastic repositories

* Add requests log repository and  refactor another

* Change repository interface to Promise return types

* Add support for the query predicates

* Add expiration time for all entities

* Fix a stub response body

* Add validation for services with same poxy host

* Add validation for the stub predicate collisions

* Add a uid dependency

* Remove bug with const moment date

* Reduce number of clusters

* Fix "now is not defined"

+++ integration_tests/helpers/utils.js
@@ -8,36 +8,42 @@ module.exports = {
     return location.split("/").slice(-1)[0];
   },
   flushDB: function () {
-    return httpCLient.delete('http://localhost:9200/hermet').then(function (response) {
-      return httpCLient.put('http://localhost:9200/hermet', {
-        "mappings": {
-          "service": {
-            "properties": {
-              "proxyHost": {
-                "type": "keyword"
+    if (config.standalone) {
+      return hermetApiClient.post('/flush').then(function (result) {
+        return result;
+      });
+    } else {
+      return httpCLient.delete('http://localhost:9200/hermet').then(function (response) {
+        return httpCLient.put('http://localhost:9200/hermet', {
+          "mappings": {
+            "service": {
+              "properties": {
+                "proxyHost": {
+                  "type": "keyword"
+                }
               }
-            }
-          },
-          "stub": {
-            "properties": {
-              "response": {
-                "enabled": false
-              },
-              "predicates": {
-                "enabled": false
-              },
-              "serviceId": {
-                "type":  "keyword"
-              },
-              "sessionId": {
-                "type":  "keyword"
+            },
+            "stub": {
+              "properties": {
+                "response": {
+                  "enabled": false
+                },
+                "predicates": {
+                  "enabled": false
+                },
+                "serviceId": {
+                  "type":  "keyword"
+                },
+                "sessionId": {
+                  "type":  "keyword"
+                }
               }
             }
           }
-        }
-      }).then(function (response) {
-        return response;
+        }).then(function (response) {
+          return response;
+        });
       });
-    })
+    }
   }
 };
